requirements.txt
Appium-Python-Client>=3.2.1
selenium>=4.22.0
pytest>=8.2.0
pytest-html>=4.1.1
requests>=2.32.3
